Replaced CircuitBreaker.allowsExecution with permitting
Presently CircuitBreaker tries to track how many executions are in progress. Standalone executions use two methods for this: allowsExecution followed by preExecute. This commit reverses the permitting process, so rather than tracking how many executions are in progress we simply track how many executions are permitted, and provide a single method to acquire a permit if possible and return whether execution is permitted.