Add checked predicates, clarify the meaning of failure vs exception
This commit adds checked predicates and bi-predicates to config builders. This extra lenience will make it easier to build policies, and always should have been supported since exceptions from predicates are ignored anyways.

This commit also clarifies the meaning of "failure" vs "exception". Unfortunately, "failure" has been very overloaded in Failsafe to mean an exception or something that a policy actually considers a failure, which may or may not be a particular exception.

This commit deprecates many of the fields that record or get "failures" where failure is really meant to mean an exception, and adds new methods with "exception" naming.